requirements.txt
anyio==3.6.2
click==8.1.3
cssselect2==0.7.0
fastapi==0.89.1
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
iso3166==2.1.1
lxml==4.9.2
Pillow==9.4.0
pydantic==1.10.4
python-barcode==0.14.0
python-decouple==3.7
python-dotenv==0.21.0
python-stdnum==1.18
PyYAML==6.0
qrbill==0.7.0
qrcode==7.3.1
reportlab==3.6.12
sniffio==1.3.0
starlette==0.22.0
svglib==1.5.1
svgwrite==1.4.3
tinycss2==1.2.1
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
zipp==3.11.0
